Refactor CLI version check
